auto exposure

+++ shaders/lib/common/config.glsl
@@ -10,7 +10,7 @@ const int colortex7Format = RGB16F;
 const int colortex8Format = RGBA16F;
 const int colortex9Format = R11F_G11F_B10F;
 const int colortex10Format = RGBA16;
-const int colortex12Format = RGBA16;
+const int colortex12Format = RGBA16F;
 
 const int colortex16Format = RGB16UI;
 const int colortex17Format = RG32UI;
